Vector Representation: Complex Polynomial

Computes complex polynomials from a list of persistence diagrams. The persistence diagram points are seen as the roots of some complex polynomial, whose coefficients are returned in a complex vector. See https://link.springer.com/chapter/10.1007%2F978-3-319-23231-7_27 for more details.

Author

Mathieu Carrière

Super classes

rgudhi::PythonClass -> rgudhi::SKLearnClass -> rgudhi::VectorRepresentationStep -> ComplexPolynomial

Methods

Public methods

ComplexPolynomial$new()

ComplexPolynomial$clone()

Inherited methods

rgudhi::PythonClass$get_python_class()

rgudhi::PythonClass$set_python_class()

rgudhi::SKLearnClass$get_params()

rgudhi::SKLearnClass$set_params()

rgudhi::VectorRepresentationStep$apply()

rgudhi::VectorRepresentationStep$fit()

rgudhi::VectorRepresentationStep$fit_transform()

rgudhi::VectorRepresentationStep$transform()

Method new()

The ComplexPolynomial constructor.

Usage

ComplexPolynomial$new(polynomial_type = c("R", "S", "T"), threshold = 10)

Arguments

polynomial_type

A string specifying the Type of complex polynomial that is going to be computed (explained in https://link.springer.com/chapter/10.1007%2F978-3-319-23231-7_27). Choices are c("R", "S", "T"). Defaults to "R".

threshold

An integer value specifying the number of coefficients. This is the dimension of the complex vector of coefficients, i.e. the number of coefficients corresponding to the largest degree terms of the polynomial. If -1, this threshold is computed from the list of persistence diagrams by considering the one with the largest number of points and using the dimension of its corresponding complex vector of coefficients as threshold. Defaults to 10L.

Returns

An object of class ComplexPolynomial.

Method clone()

The objects of this class are cloneable with this method.

Usage

ComplexPolynomial$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

X <- seq_circle(10)
ac <- AlphaComplex$new(points = X)
st <- ac$create_simplex_tree()
dgm <- st$compute_persistence()$persistence_intervals_in_dimension(0)
ds <- DiagramSelector$new(use = TRUE)
dgm <- ds$apply(dgm)
cp <- ComplexPolynomial$new()
cp$apply(dgm)
#> # A tibble: 10 × 2
#>     Grid Value                      
#>    <int> <cpl[1d]>                  
#>  1     1  0.000000e+00-8.594235e-01i
#>  2     2 -3.282706e-01+0.000000e+00i
#>  3     3  0.000000e+00+7.314312e-02i
#>  4     4  1.047682e-02+0.000000e+00i
#>  5     5  0.000000e+00-1.000447e-03i
#>  6     6 -6.368947e-05+0.000000e+00i
#>  7     7  0.000000e+00+2.606487e-06i
#>  8     8  6.222435e-08+0.000000e+00i
#>  9     9  0.000000e+00-6.602107e-10i
#> 10    10  0.000000e+00+0.000000e+00i
cp$fit_transform(list(dgm))
#> # A tibble: 10 × 2
#>     Grid Value    
#>    <int> <list>   
#>  1     1 <cpl [1]>
#>  2     2 <cpl [1]>
#>  3     3 <cpl [1]>
#>  4     4 <cpl [1]>
#>  5     5 <cpl [1]>
#>  6     6 <cpl [1]>
#>  7     7 <cpl [1]>
#>  8     8 <cpl [1]>
#>  9     9 <cpl [1]>
#> 10    10 <cpl [1]>