R6 Class for Cover Complex

Nerves and Graph Induced Complexes are cover complexes, i.e. simplicial complexes that provably contain topological information about the input data. They can be computed with a cover of the data, that comes i.e. from the pre-image of a family of intervals covering the image of a scalar-valued function defined on the data.

Details

Cover complex data structure. The data structure is a simplicial complex, representing a Graph Induced simplicial Complex (GIC) or a Nerve, and whose simplices are computed with a cover \(C\) of a point cloud \(P\), which often comes from the pre-images of intervals covering the image of a function \(f\) defined on \(P\). These intervals are parameterised by their resolution (either their length or their number) and their gain (percentage of overlap). To compute a GIC, one also needs a graph \(G\) built on top of \(P\), whose cliques with vertices belonging to different elements of \(C\) correspond to the simplices of the GIC.

See also

Author

Mathieu Carrière

Methods

Public methods

CoverComplex$new()

CoverComplex$compute_PD()

CoverComplex$compute_confidence_level_from_distance()

CoverComplex$compute_distance_from_confidence_level()

CoverComplex$compute_distribution()

CoverComplex$compute_p_value()

CoverComplex$create_simplex_tree()

CoverComplex$find_simplices()

CoverComplex$plot_dot()

CoverComplex$clone()

Method new()

CoverComplex constructor.

Usage

CoverComplex$new()

Returns

A CoverComplex object storing the Cover complex.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc
}

Method compute_PD()

Computes the extended persistence diagram of the complex.

Usage

CoverComplex$compute_PD()

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_PD()
}

Method compute_confidence_level_from_distance()

Computes the confidence level of a specific Bottleneck distance threshold.

Usage

CoverComplex$compute_confidence_level_from_distance(distance_threshold)

Arguments

distance_threshold

A numeric value specifying the desired Bottleneck distance threshold.

Returns

A numeric value storing the confidence level corresponding to the input threshold on the Bottleneck distance.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_confidence_level_from_distance(distance_threshold = 0.1)
}

Method compute_distance_from_confidence_level()

Computes the bottleneck distance threshold corresponding to a specific confidence level.

Usage

CoverComplex$compute_distance_from_confidence_level(confidence_level)

Arguments

confidence_level

A numeric value specifying the desired confidence level.

Returns

A numeric value storing the threshold on the Bottleneck distance corresponding to the input confidence level.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distance_from_confidence_level(confidence_level = 0.1)
}

Method compute_distribution()

Computes the distribution of distances via bootstrap.

Usage

CoverComplex$compute_distribution(N = 100L)

Arguments

N

An integer value specifying the number of iterations. Defaults to 100L.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distribution()
}

Method compute_p_value()

Computes the p-value, i.e. the opposite of the confidence level of the largest bottleneck distance preserving the points in the persistence diagram of the output simplicial complex.

Usage

CoverComplex$compute_p_value()

Returns

A numeric value storing the desired p-value.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_p_value()
}

Method create_simplex_tree()

Usage

CoverComplex$create_simplex_tree()

Returns

A SimplexTree object storing the simplex tree created from the Cover complex.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  st <- cc$create_simplex_tree()
}

Method find_simplices()

Computes the simplices of the simplicial complex.

Usage

CoverComplex$find_simplices()

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$find_simplices()
}

Method plot_dot()

Creates a .dot file called _sc.dot for neato (part of the graphviz Python package) once the simplicial complex is computed to get a visualization of its 1-skeleton in a .pdf file.

Usage

CoverComplex$plot_dot(dir = getwd())

Arguments

dir

A character string specifying the path to a directory into which the .dot file will be saved. Defaults to current working directory retrieved via getwd().

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  withr::with_tempdir({
    cc$plot_dot()
  })
}

Method clone()

The objects of this class are cloneable with this method.

Usage

CoverComplex$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `CoverComplex$new`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc
}
#> <CoverComplex>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     compute_PD: function () 
#>     compute_confidence_level_from_distance: function (distance_threshold) 
#>     compute_distance_from_confidence_level: function (confidence_level) 
#>     compute_distribution: function (N = 100L) 
#>     compute_p_value: function () 
#>     create_simplex_tree: function () 
#>     find_simplices: function () 
#>     initialize: function () 
#>     plot_dot: function (dir = getwd()) 
#>   Private:
#>     m_ComputedBootstrapDistribution: FALSE
#>     m_ComputedPersistenceDiagram: FALSE
#>     m_ComputedSimplexTree: FALSE
#>     m_PythonClass: gudhi.nerve_gic.CoverComplex, python.builtin.object

## ------------------------------------------------
## Method `CoverComplex$compute_PD`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_PD()
}

## ------------------------------------------------
## Method `CoverComplex$compute_confidence_level_from_distance`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_confidence_level_from_distance(distance_threshold = 0.1)
}
#> [1] 1

## ------------------------------------------------
## Method `CoverComplex$compute_distance_from_confidence_level`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distance_from_confidence_level(confidence_level = 0.1)
}

## ------------------------------------------------
## Method `CoverComplex$compute_distribution`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distribution()
}

## ------------------------------------------------
## Method `CoverComplex$compute_p_value`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_p_value()
}
#> [1] 0

## ------------------------------------------------
## Method `CoverComplex$create_simplex_tree`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  st <- cc$create_simplex_tree()
}

## ------------------------------------------------
## Method `CoverComplex$find_simplices`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$find_simplices()
}

## ------------------------------------------------
## Method `CoverComplex$plot_dot`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  withr::with_tempdir({
    cc$plot_dot()
  })
}