R6 Class for Strong Witness Complex

A Witness complex \(\mathrm{Wit}(W,L)\) is a simplicial complex defined on two sets of points in \(\mathbb{R}^D\). The data structure is described in Boissonnat and Maria (2014) .

Details

The class constructs a (strong) witness complex for a given table of nearest landmarks with respect to witnesses.

Super class

rgudhi::WitnessComplex -> StrongWitnessComplex

Methods

Public methods

StrongWitnessComplex$new()

StrongWitnessComplex$clone()

Inherited methods

rgudhi::WitnessComplex$create_simplex_tree()

Method new()

StrongWitnessComplex constructor.

Usage

StrongWitnessComplex$new(nearest_landmark_table)

Arguments

nearest_landmark_table

A list of tibbles specifying for each witness w, the ordered list of nearest landmarks with id in column nearest_landmark and distance to w in column distance.

Returns

A StrongWitnessComplex object storing the strong Witness complex.

Examples

withr::with_seed(1234, {
  l <- list(
    tibble::tibble(
      nearest_landmark = sample.int(10),
      distance = sort(rexp(10))
    ),
    tibble::tibble(
      nearest_landmark = sample.int(10),
      distance = sort(rexp(10))
    )
  )
})
if (reticulate::py_module_available("gudhi")) {
  wc <- StrongWitnessComplex$new(nearest_landmark_table = l)
  wc
}

Method clone()

The objects of this class are cloneable with this method.

Usage

StrongWitnessComplex$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `StrongWitnessComplex$new`
## ------------------------------------------------

withr::with_seed(1234, {
  l <- list(
    tibble::tibble(
      nearest_landmark = sample.int(10),
      distance = sort(rexp(10))
    ),
    tibble::tibble(
      nearest_landmark = sample.int(10),
      distance = sort(rexp(10))
    )
  )
})
if (reticulate::py_module_available("gudhi")) {
  wc <- StrongWitnessComplex$new(nearest_landmark_table = l)
  wc
}
#> <StrongWitnessComplex>
#>   Inherits from: <WitnessComplex>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     create_simplex_tree: function (max_alpha_square = Inf) 
#>     initialize: function (nearest_landmark_table) 
#>   Private:
#>     m_ComputedSimplexTree: FALSE
#>     m_PythonClass: gudhi.strong_witness_complex.StrongWitnessComplex, python.builtin.object