Vector Representation: Topological Vector

Computes topological vectors from a list of persistence diagrams. The topological vector associated to a persistence diagram is the sorted vector of a slight modification of the pairwise distances between the persistence diagram points. See https://diglib.eg.org/handle/10.1111/cgf12692 for more details.

Author

Mathieu Carrière

Super classes

rgudhi::PythonClass -> rgudhi::SKLearnClass -> rgudhi::VectorRepresentationStep -> TopologicalVector

Methods

Public methods

TopologicalVector$new()

TopologicalVector$clone()

Inherited methods

rgudhi::PythonClass$get_python_class()

rgudhi::PythonClass$set_python_class()

rgudhi::SKLearnClass$get_params()

rgudhi::SKLearnClass$set_params()

rgudhi::VectorRepresentationStep$apply()

rgudhi::VectorRepresentationStep$fit()

rgudhi::VectorRepresentationStep$fit_transform()

rgudhi::VectorRepresentationStep$transform()

Method new()

The TopologicalVector constructor.

Usage

TopologicalVector$new(threshold = 10)

Arguments

threshold

An integer value specifying the number of distances to keep. Defaults to 10L. This is the dimension of the topological vector. If -1, this threshold is computed from the list of persistence diagrams by considering the one with the largest number of points and using the dimension of its corresponding topological vector as threshold.

Returns

An object of class TopologicalVector.

Examples

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n + 1)[1:n],
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  ac <- AlphaComplex$new(points = X)
  st <- ac$create_simplex_tree()
  dgm <- st$compute_persistence()$persistence_intervals_in_dimension(0)
  ds <- DiagramSelector$new(use = TRUE)
  dgm <- ds$apply(dgm)
  tv <- TopologicalVector$new()
  tv$apply(dgm)
  tv$fit_transform(list(dgm))
}

Method clone()

The objects of this class are cloneable with this method.

Usage

TopologicalVector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `TopologicalVector$new`
## ------------------------------------------------

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n + 1)[1:n],
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  ac <- AlphaComplex$new(points = X)
  st <- ac$create_simplex_tree()
  dgm <- st$compute_persistence()$persistence_intervals_in_dimension(0)
  ds <- DiagramSelector$new(use = TRUE)
  dgm <- ds$apply(dgm)
  tv <- TopologicalVector$new()
  tv$apply(dgm)
  tv$fit_transform(list(dgm))
}