R6 Class for Cover Complex

Nerves and Graph Induced Complexes are cover complexes, i.e. simplicial complexes that provably contain topological information about the input data. They can be computed with a cover of the data, that comes i.e. from the pre-image of a family of intervals covering the image of a scalar-valued function defined on the data.

Details

Cover complex data structure. The data structure is a simplicial complex, representing a Graph Induced simplicial Complex (GIC) or a Nerve, and whose simplices are computed with a cover \(C\) of a point cloud \(P\), which often comes from the pre-images of intervals covering the image of a function \(f\) defined on \(P\). These intervals are parameterised by their resolution (either their length or their number) and their gain (percentage of overlap). To compute a GIC, one also needs a graph \(G\) built on top of \(P\), whose cliques with vertices belonging to different elements of \(C\) correspond to the simplices of the GIC.

See also

Author

Mathieu Carrière

Methods

Public methods

CoverComplex$new()

CoverComplex$compute_PD()

CoverComplex$compute_confidence_level_from_distance()

CoverComplex$compute_distance_from_confidence_level()

CoverComplex$compute_distribution()

CoverComplex$compute_p_value()

CoverComplex$create_simplex_tree()

CoverComplex$find_simplices()

CoverComplex$plot_dot()

CoverComplex$plot_off()

CoverComplex$read_point_cloud()

CoverComplex$set_automatic_resolution()

CoverComplex$set_color_from_coordinate()

CoverComplex$set_color_from_file()

CoverComplex$set_color_from_range()

CoverComplex$set_cover_from_Voronoi()

CoverComplex$set_cover_from_file()

CoverComplex$set_cover_from_function()

CoverComplex$set_distances_from_range()

CoverComplex$set_function_from_coordinate()

CoverComplex$set_function_from_file()

CoverComplex$set_function_from_range()

CoverComplex$set_gain()

CoverComplex$set_graph_from_OFF()

CoverComplex$set_graph_from_automatic_rips()

CoverComplex$set_graph_from_file()

CoverComplex$set_graph_from_rips()

CoverComplex$set_mask()

CoverComplex$set_point_cloud_from_range()

CoverComplex$set_resolution_with_interval_length()

CoverComplex$set_resolution_with_interval_number()

CoverComplex$set_subsampling()

CoverComplex$set_verbose()

CoverComplex$subpopulation()

CoverComplex$write_info()

CoverComplex$clone()

Method new()

CoverComplex constructor.

Usage

CoverComplex$new(type)

Arguments

type

A character string specifying the type of output simplicial complex. Can be either "GIC" or "Nerve".

Returns

A CoverComplex object storing the Cover complex.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new(type = "GIC")
  cc
}

Method compute_PD()

Computes the extended persistence diagram of the complex.

Usage

CoverComplex$compute_PD()

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_PD()
}

Method compute_confidence_level_from_distance()

Computes the confidence level of a specific Bottleneck distance threshold.

Usage

CoverComplex$compute_confidence_level_from_distance(distance_threshold)

Arguments

distance_threshold

A numeric value specifying the desired Bottleneck distance threshold.

Returns

A numeric value storing the confidence level corresponding to the input threshold on the Bottleneck distance.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_confidence_level_from_distance(distance_threshold = 0.1)
}

Method compute_distance_from_confidence_level()

Computes the bottleneck distance threshold corresponding to a specific confidence level.

Usage

CoverComplex$compute_distance_from_confidence_level(confidence_level)

Arguments

confidence_level

A numeric value specifying the desired confidence level.

Returns

A numeric value storing the threshold on the Bottleneck distance corresponding to the input confidence level.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distance_from_confidence_level(confidence_level = 0.1)
}

Method compute_distribution()

Computes the distribution of distances via bootstrap.

Usage

CoverComplex$compute_distribution(N = 100L, dir = getwd())

Arguments

N

An integer value specifying the number of iterations. Defaults to 100L.

dir

A character string specifying the path to a directory into which the .off file will be saved. Defaults to current working directory retrieved via getwd().

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distribution()
}

Method compute_p_value()

Computes the p-value, i.e. the opposite of the confidence level of the largest bottleneck distance preserving the points in the persistence diagram of the output simplicial complex.

Usage

CoverComplex$compute_p_value()

Returns

A numeric value storing the desired p-value.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_p_value()
}

Method create_simplex_tree()

Usage

CoverComplex$create_simplex_tree()

Returns

A SimplexTree object storing the simplex tree created from the Cover complex.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  st <- cc$create_simplex_tree()
}

Method find_simplices()

Computes the simplices of the simplicial complex.

Usage

CoverComplex$find_simplices()

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$find_simplices()
}

Method plot_dot()

Creates a .dot file called _sc.dot for neato (part of the graphviz Python package) once the simplicial complex is computed to get a visualization of its 1-skeleton in a .pdf file.

Usage

CoverComplex$plot_dot(dir = getwd())

Arguments

dir

A character string specifying the path to a directory into which the .dot file will be saved. Defaults to current working directory retrieved via getwd().

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  withr::with_tempdir({
    cc$plot_dot()
  })
}

Method plot_off()

Creates a .off file called _sc.off for 3D visualization, which contains the 2-skeleton of the GIC. This function assumes that the cover has been computed with Voronoi. If data points are in 1D or 2D, the remaining coordinates of the points embedded in 3D are set to 0.

Usage

CoverComplex$plot_off(dir = getwd())

Arguments

dir

A character string specifying the path to a directory into which the .off file will be saved. Defaults to current working directory retrieved via getwd().

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  withr::with_tempdir({
    cc$plot_dot()
  })
}

Method read_point_cloud()

Reads and stores the input point cloud from a .(n)OFF file.

Usage

CoverComplex$read_point_cloud(off_file, chainable = TRUE)

Arguments

off_file

A character string specifying the location of the .(n)OFF file to read the point cloud from.

chainable

A boolean specyfing whether the method should be chainable in which case it returns invisibly the class itself. Defaults to TRUE.

Returns

The updated CoverComplex class itself invisibly if chainable = TRUE or a boolean storing the read file status.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  url <- "https://raw.githubusercontent.com/GUDHI/TDA-tutorial/master/datasets/tore3D_1307.off"
  cc$read_point_cloud(url)
}

Method set_automatic_resolution()

Computes the optimal length of intervals (i.e. the smallest interval length avoiding discretization artifacts - see Carriere et al. (2018) ) for a functional cover.

References

Carriere M, Michel B, Oudot S (2018). “Statistical analysis and parameter selection for mapper.” The Journal of Machine Learning Research, 19(1), 478--516.

Usage

CoverComplex$set_automatic_resolution(chainable = TRUE)

Arguments

chainable

A boolean specyfing whether the method should be chainable in which case it returns invisibly the class itself. Defaults to TRUE.

Returns

The updated CoverComplex class itself invisibly if chainable = TRUE or a numeric value storing the resolution interval length used to compute the cover.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_automatic_resolution()
}

Method set_color_from_coordinate()

Computes the function used to color the nodes of the simplicial complex from the k-th coordinate.

Usage

CoverComplex$set_color_from_coordinate(k = 0)

Arguments

k

An integer value specifying the coordinate to use (start at 0). Defaults to 0L.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_color_from_coordinate()
}

Method set_color_from_file()

Computes the function used to color the nodes of the simplicial complex from a file containing the function values.

Usage

CoverComplex$set_color_from_file(color_file_name)

Arguments

color_file_name

A character string specifying the name of the input color file.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_color_from_file()
}

Method set_color_from_range()

Computes the function used to color the nodes of the simplicial complex from a vector stored in memory.

Usage

CoverComplex$set_color_from_range(color_values)

Arguments

color_values

A numeric vector specifying the input vector of values.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_color_from_range(seq(0, 1, len = 100))
}

Method set_cover_from_Voronoi()

Creates the cover \(C\) from the Voronoï cells of a subsampling of the point cloud.

Usage

CoverComplex$set_cover_from_Voronoi(m = 100L)

Arguments

m

An integer value specifying the number of points in the subsample. Defaults to 100L.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_cover_from_Voronoi()
}

Method set_cover_from_file()

Creates the cover \(C\) from a file containing the cover elements of each point (the order has to be the same as in the input file!).

Usage

CoverComplex$set_cover_from_file(cover_file_name)

Arguments

cover_file_name

A character string specifying the path to the input cover file.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_cover_from_file()
}

Method set_cover_from_function()

Creates a cover \(C\) from the pre-images of the function \(f\).

Usage

CoverComplex$set_cover_from_function()

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_cover_from_function()
}

Method set_distances_from_range()

Reads and stores the input distance matrix from a vector stored in memory.

Usage

CoverComplex$set_distances_from_range(distance_matrix)

Arguments

distance_matrix

A numeric matrix specifying the distance matrix.

Returns

The updated CoverComplex class itself invisibly.

Examples

D <- dist(iris[, -5])
if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_distances_from_range(D)
}

Method set_function_from_coordinate()

Creates the function \(f\) from the \(k\)-th coordinate of the point cloud.

Usage

CoverComplex$set_function_from_coordinate(k = 0L)

Arguments

k

An integer value specifying the coordinate to use (starts at 0L). Defaults to 0L.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_function_from_coordinate()
}

Method set_function_from_file()

Creates the function \(f\) from a file containing the function values.

Usage

CoverComplex$set_function_from_file(func_file_name)

Arguments

func_file_name

A character string specifying the path to the input function file.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_function_from_file()
}

Method set_function_from_range()

Creates the function \(f\) from a vector stored in memory.

Usage

CoverComplex$set_function_from_range(function_values)

Arguments

function_values

A numeric vector specifying the function values to be used.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_function_from_range(seq(0, 1, len = 100))
}

Method set_gain()

Sets a gain from a value stored in memory.

Usage

CoverComplex$set_gain(g = 0.3)

Arguments

g

A numeric value specifying the gain. Defaults to 0.3.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_gain()
}

Method set_graph_from_OFF()

Creates a graph \(G\) from the triangulation given by the input .off file.

Usage

CoverComplex$set_graph_from_OFF()

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_graph_from_OFF()
}

Method set_graph_from_automatic_rips()

Creates a graph \(G\) from a Rips complex whose threshold value is automatically tuned with subsampling - see Carriere et al. (2018) .

References

Carriere M, Michel B, Oudot S (2018). “Statistical analysis and parameter selection for mapper.” The Journal of Machine Learning Research, 19(1), 478--516.

Usage

CoverComplex$set_graph_from_automatic_rips(N = 100L, chainable = TRUE)

Arguments

N

An integer value specifying the number of subsampling iterations. Defaults to 100L but there is no guarantee on how to choose it.

chainable

A boolean specyfing whether the method should be chainable in which case it returns invisibly the class itself. Defaults to TRUE.

Returns

The updated CoverComplex class itself invisibly if chainable = TRUE or a numeric value storing the delta threshold used for computing the Rips complex..

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_graph_from_automatic_rips(chainable = FALSE)
}

Method set_graph_from_file()

Creates a graph \(G\) from a file containing the edges.

Usage

CoverComplex$set_graph_from_file(graph_file_name)

Arguments

graph_file_name

A character string specifying the path to the input graph file. The graph file contains one edge per line, each edge being represented by the IDs of its two nodes.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_graph_from_file()
}

Method set_graph_from_rips()

Creates a graph \(G\) from a Rips complex.

Usage

CoverComplex$set_graph_from_rips(threshold)

Arguments

threshold

A numeric value specifying the threshold value for the Rips complex.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_graph_from_rips(0.1)
}

Method set_mask()

Sets the mask, which is a threshold integer such that nodes in the complex that contain a number of data points which is less than or equal to this threshold are not displayed.

Usage

CoverComplex$set_mask(nodemask)

Arguments

nodemask

A numeric value specifying the threshold value for generating the mask.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_mask(0.5)
}

Method set_point_cloud_from_range()

Reads and stores the input point cloud from a vector stored in memory.

Usage

CoverComplex$set_point_cloud_from_range(cloud)

Arguments

cloud

A numeric matrix specifying the coordinates of the point cloud.

Returns

The updated CoverComplex class itself invisibly.

Examples

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n),
  function(.x) c(cos(.x), sin(.x))
)
X <- Reduce(rbind, X, init = numeric())
if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_point_cloud_from_range(X)
}

Method set_resolution_with_interval_length()

Sets a length of intervals from a value stored in memory.

Usage

CoverComplex$set_resolution_with_interval_length(resolution)

Arguments

resolution

A numeric value specifying the length of intervals.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_resolution_with_interval_length(1)
}

Method set_resolution_with_interval_number()

Sets a number of intervals from a value stored in memory.

Usage

CoverComplex$set_resolution_with_interval_number(resolution)

Arguments

resolution

An integer value specifying the number of intervals.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_resolution_with_interval_number(100L)
}

Method set_subsampling()

Sets the constants used to subsample the data set. These constants are explained in Carriere et al. (2018) .

References

Carriere M, Michel B, Oudot S (2018). “Statistical analysis and parameter selection for mapper.” The Journal of Machine Learning Research, 19(1), 478--516.

Usage

CoverComplex$set_subsampling(constant, power)

Arguments

constant

A numeric value specifying the subsampling constant.

power

A numeric value specifying the subsampling power.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_subsampling(constant = 0, power = 1)
}

Method set_verbose()

Specifies whether the program should display information or not.

Usage

CoverComplex$set_verbose(verbose = FALSE)

Arguments

verbose

A boolean specifying whether to display information. Defaults to FALSE.

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_verbose(FALSE)
}

Method subpopulation()

Returns the data subset corresponding to a specific node of the created complex.

Usage

CoverComplex$subpopulation(node_id)

Arguments

node_id

An integer value specifying the ID of the desired node.

Returns

An integer vector storing the IDs of the data points at the input node.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$subpopulation(0)
}

Method write_info()

Creates a .txt file called _sc.txt describing the 1-skeleton, which can then be plotted with e.g. KeplerMapper.

Usage

CoverComplex$write_info(dir = getwd())

Arguments

dir

A character string specifying the path to a directory into which the .txt file will be saved. Defaults to current working directory retrieved via getwd().

Returns

The updated CoverComplex class itself invisibly.

Examples

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$write_info()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

CoverComplex$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `CoverComplex$new`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new(type = "GIC")
  cc
}
#> <CoverComplex>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     compute_PD: function () 
#>     compute_confidence_level_from_distance: function (distance_threshold) 
#>     compute_distance_from_confidence_level: function (confidence_level) 
#>     compute_distribution: function (N = 100L, dir = getwd()) 
#>     compute_p_value: function () 
#>     create_simplex_tree: function () 
#>     find_simplices: function () 
#>     initialize: function (type) 
#>     plot_dot: function (dir = getwd()) 
#>     plot_off: function (dir = getwd()) 
#>     read_point_cloud: function (off_file, chainable = TRUE) 
#>     set_automatic_resolution: function (chainable = TRUE) 
#>     set_color_from_coordinate: function (k = 0) 
#>     set_color_from_file: function (color_file_name) 
#>     set_color_from_range: function (color_values) 
#>     set_cover_from_Voronoi: function (m = 100L) 
#>     set_cover_from_file: function (cover_file_name) 
#>     set_cover_from_function: function () 
#>     set_distances_from_range: function (distance_matrix) 
#>     set_function_from_coordinate: function (k = 0L) 
#>     set_function_from_file: function (func_file_name) 
#>     set_function_from_range: function (function_values) 
#>     set_gain: function (g = 0.3) 
#>     set_graph_from_OFF: function () 
#>     set_graph_from_automatic_rips: function (N = 100L, chainable = TRUE) 
#>     set_graph_from_file: function (graph_file_name) 
#>     set_graph_from_rips: function (threshold) 
#>     set_mask: function (nodemask) 
#>     set_point_cloud_from_range: function (cloud) 
#>     set_resolution_with_interval_length: function (resolution) 
#>     set_resolution_with_interval_number: function (resolution) 
#>     set_subsampling: function (constant, power) 
#>     set_verbose: function (verbose = FALSE) 
#>     subpopulation: function (node_id) 
#>     write_info: function (dir = getwd()) 
#>   Private:
#>     m_ComputedBootstrapDistribution: FALSE
#>     m_ComputedPersistenceDiagram: FALSE
#>     m_ComputedSimplexTree: FALSE
#>     m_ComputedSimplicies: FALSE
#>     m_IsCoverDefined: FALSE
#>     m_IsCoverDefinedFromFunction: FALSE
#>     m_IsFunctionDefined: FALSE
#>     m_IsGraphDefined: FALSE
#>     m_IsGraphDefinedFromAutomaticRips: FALSE
#>     m_IsPointCloudDefined: FALSE
#>     m_IsPointCloudDefinedFromOFF: FALSE
#>     m_IsResolutionDefined: FALSE
#>     m_PythonClass: gudhi.nerve_gic.CoverComplex, python.builtin.object

## ------------------------------------------------
## Method `CoverComplex$compute_PD`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_PD()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$compute_confidence_level_from_distance`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_confidence_level_from_distance(distance_threshold = 0.1)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$compute_distance_from_confidence_level`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distance_from_confidence_level(confidence_level = 0.1)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$compute_distribution`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$compute_distribution()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$compute_p_value`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$compute_p_value()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$create_simplex_tree`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  st <- cc$create_simplex_tree()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$find_simplices`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$find_simplices()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$plot_dot`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  withr::with_tempdir({
    cc$plot_dot()
  })
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$plot_off`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  withr::with_tempdir({
    cc$plot_dot()
  })
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$read_point_cloud`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  url <- "https://raw.githubusercontent.com/GUDHI/TDA-tutorial/master/datasets/tore3D_1307.off"
  cc$read_point_cloud(url)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_automatic_resolution`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_automatic_resolution()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_color_from_coordinate`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_color_from_coordinate()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_color_from_file`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_color_from_file()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_color_from_range`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_color_from_range(seq(0, 1, len = 100))
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_cover_from_Voronoi`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_cover_from_Voronoi()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_cover_from_file`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_cover_from_file()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_cover_from_function`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_cover_from_function()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_distances_from_range`
## ------------------------------------------------

D <- dist(iris[, -5])
if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_distances_from_range(D)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_function_from_coordinate`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_function_from_coordinate()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_function_from_file`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_function_from_file()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_function_from_range`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_function_from_range(seq(0, 1, len = 100))
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_gain`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_gain()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_graph_from_OFF`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_graph_from_OFF()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_graph_from_automatic_rips`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_graph_from_automatic_rips(chainable = FALSE)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_graph_from_file`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  # cc$set_graph_from_file()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_graph_from_rips`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_graph_from_rips(0.1)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_mask`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_mask(0.5)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_point_cloud_from_range`
## ------------------------------------------------

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n),
  function(.x) c(cos(.x), sin(.x))
)
X <- Reduce(rbind, X, init = numeric())
if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_point_cloud_from_range(X)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_resolution_with_interval_length`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_resolution_with_interval_length(1)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_resolution_with_interval_number`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_resolution_with_interval_number(100L)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_subsampling`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_subsampling(constant = 0, power = 1)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$set_verbose`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$set_verbose(FALSE)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$subpopulation`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$subpopulation(0)
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default

## ------------------------------------------------
## Method `CoverComplex$write_info`
## ------------------------------------------------

if (reticulate::py_module_available("gudhi")) {
  cc <- CoverComplex$new()
  cc$write_info()
}
#> Error in match.arg(type, choices = c("GIC", "Nerve")): argument "type" is missing, with no default