R6 Class for Periodic Cubical Complex

The PeriodicCubicalComplex class is an example of a structured complex useful in computational mathematics (specially rigorous numerics) and image analysis.

See also

Author

Pawel Dlotko

Super classes

rgudhi::PythonClass -> rgudhi::CubicalComplex -> PeriodicCubicalComplex

Methods

Public methods

PeriodicCubicalComplex$new()

PeriodicCubicalComplex$clone()

Inherited methods

rgudhi::PythonClass$get_python_class()

rgudhi::PythonClass$set_python_class()

rgudhi::CubicalComplex$betti_numbers()

rgudhi::CubicalComplex$cofaces_of_persistence_pairs()

rgudhi::CubicalComplex$compute_persistence()

rgudhi::CubicalComplex$dimension()

rgudhi::CubicalComplex$num_simplices()

rgudhi::CubicalComplex$persistence()

rgudhi::CubicalComplex$persistence_intervals_in_dimension()

rgudhi::CubicalComplex$persistent_betti_numbers()

Method new()

Constructor from either top_dimensional_cells (and possibly dimensions) or from a Perseus-style file name.

Usage

PeriodicCubicalComplex$new(
  perseus_file,
  top_dimensional_cells,
  periodic_dimensions,
  dimensions = NULL,
  py_class = NULL
)

Arguments

perseus_file

A character string specifying the path to a Perseus-style file name.

top_dimensional_cells

Either a numeric vector (in which case, dimensions should be provided as well) or a multidimensional array specifying cell filtration values.

periodic_dimensions

A logical vector specifying the periodicity value of the top dimensional cells.

dimensions

An integer vector specifying the number of top dimensional cells. Defaults to NULL.

py_class

An existing PeriodicCubicalComplex Python class. Defaults to NULL which uses the Python class constructor instead.

Returns

A new PeriodicCubicalComplex object.

Method clone()

The objects of this class are cloneable with this method.

Usage

PeriodicCubicalComplex$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

n <- 10
X <- cbind(seq(0, 1, len = n), seq(0, 1, len = n))
pcc <- PeriodicCubicalComplex$new(
  top_dimensional_cells = X,
  periodic_dimensions = c(TRUE, FALSE)
)
pcc
#> <PeriodicCubicalComplex>
#>   Inherits from: <CubicalComplex>
#>   Public:
#>     betti_numbers: function () 
#>     clone: function (deep = FALSE) 
#>     cofaces_of_persistence_pairs: function () 
#>     compute_persistence: function (homology_coeff_field = 11, min_persistence = 0) 
#>     dimension: function () 
#>     get_python_class: function () 
#>     initialize: function (perseus_file, top_dimensional_cells, periodic_dimensions, 
#>     num_simplices: function () 
#>     persistence: function (homology_coeff_field = 11, min_persistence = 0) 
#>     persistence_intervals_in_dimension: function (dimension) 
#>     persistent_betti_numbers: function (from_value, to_value) 
#>     set_python_class: function (py_class) 
#>   Private:
#>     m_ComputedPersistence: FALSE
#>     python_class: gudhi.periodic_cubical_complex.PeriodicCubicalComplex, python.builtin.object