R6 Class for Rips Complex

The data structure is a one skeleton graph, or Rips graph, containing edges when the edge length is less or equal to a given threshold. Edge length is computed from a user given point cloud with a given distance function, or a distance matrix.

See also

Author

Clément Maria, Pawel Dlotko, Vincent Rouvreau, Marc Glisse, Yuichi Ike

Super class

rgudhi::PythonClass -> RipsComplex

Methods

Public methods

RipsComplex$new()

RipsComplex$create_simplex_tree()

RipsComplex$clone()

Inherited methods

rgudhi::PythonClass$get_python_class()

rgudhi::PythonClass$set_python_class()

Method new()

RipsComplex constructor.

Usage

RipsComplex$new(data, max_edge_length = NULL, sparse = NULL)

Arguments

data

Either a n x d matrix or a length-n list of d-dimensional vectors or a distance matrix stored as a dist object.

max_edge_length

A numeric value specifying the Rips value.

sparse

A numeric value specifying the approximation parameter epsilon for buidling a sparse Rips complex. Defaults to NULL which builds an exact Rips complex.

Returns

A RipsComplex object storing the Rips complex.

Examples

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n),
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  rc1 <- RipsComplex$new(data = X, max_edge_length = 1)
}
Xm <- Reduce(rbind, X, init = numeric())
if (reticulate::py_module_available("gudhi")) {
  rc2 <- RipsComplex$new(data = Xm, max_edge_length = 1)
}
D <- dist(Xm)
if (reticulate::py_module_available("gudhi")) {
  rc3 <- RipsComplex$new(data = D)
}

Method create_simplex_tree()

Usage

RipsComplex$create_simplex_tree(max_dimension)

Arguments

max_dimension

An integer value specifying the maximal dimension which the Rips complex will be expanded to.

Returns

A SimplexTree object storing the computed simplex tree.

Examples

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n),
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  rc <- RipsComplex$new(data = X, max_edge_length = 1)
  st <- rc$create_simplex_tree(1)
}

Method clone()

The objects of this class are cloneable with this method.

Usage

RipsComplex$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `RipsComplex$new`
## ------------------------------------------------

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n),
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  rc1 <- RipsComplex$new(data = X, max_edge_length = 1)
}
Xm <- Reduce(rbind, X, init = numeric())
if (reticulate::py_module_available("gudhi")) {
  rc2 <- RipsComplex$new(data = Xm, max_edge_length = 1)
}
D <- dist(Xm)
if (reticulate::py_module_available("gudhi")) {
  rc3 <- RipsComplex$new(data = D)
}

## ------------------------------------------------
## Method `RipsComplex$create_simplex_tree`
## ------------------------------------------------

n <- 10
X <- lapply(
  seq(0, 2 * pi, len = n),
  function(.x) c(cos(.x), sin(.x))
)
if (reticulate::py_module_available("gudhi")) {
  rc <- RipsComplex$new(data = X, max_edge_length = 1)
  st <- rc$create_simplex_tree(1)
}